Rename project name to CameraLab
