+++ app/build.gradle
@@ -7,11 +7,11 @@ plugins {
 }
 
 android {
-    namespace 'com.amoseui.cameraplayground'
+    namespace 'com.amoseui.cameralab'
 
     compileSdk 33
     defaultConfig {
-        applicationId "com.amoseui.cameraplayground"
+        applicationId "com.amoseui.cameralab"
         minSdk 21
         targetSdk 33
         versionCode 10
