+++ app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.CameraPlayground">
+        android:theme="@style/Theme.CameraLab">
         <activity
             android:name=".MainActivity"
             android:exported="true">
