+++ app/src/main/java/com/amoseui/cameralab/ui/main/SettingsFragment.kt
@@ -1,11 +1,11 @@
-package com.amoseui.cameraplayground.ui.main
+package com.amoseui.cameralab.ui.main
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.amoseui.cameraplayground.R
+import com.amoseui.cameralab.R
 
 class SettingsFragment : Fragment() {
