+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Camera Dev</string>
+    <string name="app_name">Camera Lab</string>
 </resources>
