+++ settings.gradle
@@ -3,8 +3,7 @@ dependencyResolutionManagement {
     repositories {
         google()
         mavenCentral()
-        jcenter() // Warning: this repository is going to shut down soon
     }
 }
-rootProject.name = "CameraPlayground"
+rootProject.name = "CameraLab"
 include ':app'
